define correctly boolean values in ldap3 env. var.
os.environ return a string for boolean values. A test is needed to store a boolean value in config. If Flask-ldap test a string value in IF statement, it always use True value. So app.config['LDAP_USE_SSL'] is always to True when we put a value (like False) in environment variable LDAP_USE_SSL.